+++ pkgs/native_assets_builder/test_data/drop_dylib_link/hook/build.dart
@@ -24,8 +24,8 @@ void main(List<String> arguments) async {
       dartBuildFiles: ['hook/build.dart'],
       linkModePreference: LinkModePreference.dynamic,
     ).run(
-      buildConfig: config,
-      buildOutput: output,
+      config: config,
+      output: output,
       logger: logger,
       linkInPackage: packageName,
     );
@@ -39,8 +39,8 @@ void main(List<String> arguments) async {
       dartBuildFiles: ['hook/build.dart'],
       linkModePreference: LinkModePreference.dynamic,
     ).run(
-      buildConfig: config,
-      buildOutput: output,
+      config: config,
+      output: output,
       logger: logger,
       linkInPackage: packageName,
     );