+++ pkgs/native_toolchain_c/lib/src/cbuilder/cbuilder.dart
@@ -10,14 +10,6 @@ import 'package:native_assets_cli/native_assets_cli.dart';
 
 import 'run_cbuilder.dart';
 
-abstract class Builder {
-  Future<void> run({
-    required BuildConfig buildConfig,
-    required BuildOutput buildOutput,
-    required Logger? logger,
-  });
-}
-
 /// A programming language that can be selected for compilation of source files.
 ///
 /// See [CBuilder.language] for more information.
@@ -200,20 +192,18 @@ class CBuilder implements Builder {
   /// Completes with an error if the build fails.
   @override
   Future<void> run({
-    required BuildConfig buildConfig,
-    required BuildOutput buildOutput,
+    required BuildConfig config,
+    required BuildOutput output,
     required Logger? logger,
     String? linkInPackage,
   }) async {
-    final outDir = buildConfig.outputDirectory;
-    final packageRoot = buildConfig.packageRoot;
+    final outDir = config.outputDirectory;
+    final packageRoot = config.packageRoot;
     await Directory.fromUri(outDir).create(recursive: true);
-    final linkMode =
-        _linkMode(linkModePreference ?? buildConfig.linkModePreference);
+    final linkMode = _linkMode(linkModePreference ?? config.linkModePreference);
     final libUri =
-        outDir.resolve(buildConfig.targetOS.libraryFileName(name, linkMode));
-    final exeUri =
-        outDir.resolve(buildConfig.targetOS.executableFileName(name));
+        outDir.resolve(config.targetOS.libraryFileName(name, linkMode));
+    final exeUri = outDir.resolve(config.targetOS.executableFileName(name));
     final sources = [
       for (final source in this.sources)
         packageRoot.resolveUri(Uri.file(source)),
@@ -225,9 +215,9 @@ class CBuilder implements Builder {
     final dartBuildFiles = [
       for (final source in this.dartBuildFiles) packageRoot.resolve(source),
     ];
-    if (!buildConfig.dryRun) {
+    if (!config.dryRun) {
       final task = RunCBuilder(
-        buildConfig: buildConfig,
+        buildConfig: config,
         logger: logger,
         sources: sources,
         includes: includes,
@@ -244,8 +234,8 @@ class CBuilder implements Builder {
         flags: flags,
         defines: {
           ...defines,
-          if (buildModeDefine) buildConfig.buildMode.name.toUpperCase(): null,
-          if (ndebugDefine && buildConfig.buildMode != BuildMode.debug)
+          if (buildModeDefine) config.buildMode.name.toUpperCase(): null,
+          if (ndebugDefine && config.buildMode != BuildMode.debug)
             'NDEBUG': null,
         },
         pic: pic,
@@ -257,22 +247,21 @@ class CBuilder implements Builder {
     }
 
     if (assetName != null) {
-      buildOutput.addAssets(
+      output.addAssets(
         [
           NativeCodeAsset(
-            package: buildConfig.packageName,
+            package: config.packageName,
             name: assetName!,
             file: libUri,
             linkMode: linkMode,
-            os: buildConfig.targetOS,
-            architecture:
-                buildConfig.dryRun ? null : buildConfig.targetArchitecture,
+            os: config.targetOS,
+            architecture: config.dryRun ? null : config.targetArchitecture,
           )
         ],
         linkInPackage: linkInPackage,
       );
     }
-    if (!buildConfig.dryRun) {
+    if (!config.dryRun) {
       final includeFiles = await Stream.fromIterable(includes)
           .asyncExpand(
             (include) => Directory(include.toFilePath())
@@ -282,7 +271,7 @@ class CBuilder implements Builder {
           )
           .toList();
 
-      buildOutput.addDependencies({
+      output.addDependencies({
         // Note: We use a Set here to deduplicate the dependencies.
         ...sources,
         ...includeFiles,