+++ pkgs/native_assets_builder/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.1-wip
+
+- Nothing yet.
+
 ## 0.7.0
 
 - Add support for `hook/link.dart` including dry runs.