+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.6.1-wip
+
+- Introduce `Builder` and `Linker` interface.
+
 ## 0.6.0
 
 - Add support for `hook/link.dart`.