+++ pkgs/native_toolchain_c/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.0-wip
+
+- Renamed parameters in `Builder.run`.
+
 ## 0.4.2
 
 - Bump `package:native_assets_cli` to 0.5.0.